Rename document_id module to articles for consistent naming
