+++ components/veritasai/articles/__init__.py
@@ -1,0 +1,3 @@
+from .dedup import generate_id
+
+__all__ = ["generate_id"]
