+++ projects/analysis-manager/pyproject.toml
@@ -24,6 +24,6 @@ includes = ["main.py", "requirements.txt"]
 [tool.polylith.bricks]
 "../../bases/veritasai/analysis_manager" = "veritasai/analysis_manager"
 "../../components/veritasai/input_validation" = "veritasai/input_validation"
-"../../components/veritasai/document_id" = "veritasai/document_id"
 "../../components/veritasai/cache" = "veritasai/cache"
 "../../components/veritasai/firebase" = "veritasai/firebase"
+"../../components/veritasai/articles" = "veritasai/articles"
